@@ -7,14 +7,12 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "@types/node": "^12.20.36",
-    "@types/react": "^17.0.34",
-    "@types/react-dom": "^17.0.11",
     "eslint-config-prettier": "^8.3.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "styled-components": "^5.3.3",
-    "typescript": "^4.4.4",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
@@ -45,13 +43,14 @@
     ]
   },
   "devDependencies": {
-    "@types/jest": "^26.0.24",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
     "@types/styled-components": "^5.1.15",
     "@typescript-eslint/eslint-plugin": "^5.3.1",
     "@typescript-eslint/parser": "^5.3.1",
     "eslint": "^8.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "prettier": "2.4.1"
+    "prettier": "2.4.1",
+    "typescript": "^4.7.4"
   }
 }
